rename modules and add method of getting all stations
